Goated Push
Barebones relay functionality is here!

When an alarm starts it makes a relay and keeps track of steps (or in the case of this dev model just times the central button was tapped) and ends the alarm when the step goal is met, Restarts step count if failed. We are getting close!